Issue #9173 - Make wrapping of ServletApiResponse easier
