+++ jetty-ee10/jetty-ee10-servlet/src/main/java/org/eclipse/jetty/ee10/servlet/ServletChannel.java
@@ -804,7 +804,7 @@ public class ServletChannel
      */
     public void onCompleted()
     {
-        ServletContextRequest.ServletApiRequest apiRequest = _servletContextRequest.getServletApiRequest();
+        ServletApiRequest apiRequest = _servletContextRequest.getServletApiRequest();
         if (LOG.isDebugEnabled())
             LOG.debug("onCompleted for {} written={}", apiRequest.getRequestURI(), getBytesWritten());
