+++ jetty-ee10/jetty-ee10-servlet/src/main/java/org/eclipse/jetty/ee10/servlet/ServletContextResponse.java
@@ -14,58 +14,35 @@
 package org.eclipse.jetty.ee10.servlet;
 
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.nio.channels.IllegalSelectorException;
-import java.util.Collection;
-import java.util.EnumSet;
 import java.util.Iterator;
 import java.util.Locale;
 import java.util.Map;
-import java.util.concurrent.CompletableFuture;
 import java.util.function.Supplier;
 
 import jakarta.servlet.ServletContext;
-import jakarta.servlet.ServletOutputStream;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.ServletResponseWrapper;
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import org.eclipse.jetty.ee10.servlet.ServletContextRequest.ServletApiRequest;
-import org.eclipse.jetty.ee10.servlet.writer.EncodingHttpWriter;
-import org.eclipse.jetty.ee10.servlet.writer.Iso88591HttpWriter;
 import org.eclipse.jetty.ee10.servlet.writer.ResponseWriter;
-import org.eclipse.jetty.ee10.servlet.writer.Utf8HttpWriter;
 import org.eclipse.jetty.http.HttpCookie;
 import org.eclipse.jetty.http.HttpField;
 import org.eclipse.jetty.http.HttpFields;
 import org.eclipse.jetty.http.HttpGenerator;
 import org.eclipse.jetty.http.HttpHeader;
 import org.eclipse.jetty.http.HttpHeaderValue;
-import org.eclipse.jetty.http.HttpScheme;
 import org.eclipse.jetty.http.HttpStatus;
-import org.eclipse.jetty.http.HttpURI;
 import org.eclipse.jetty.http.HttpVersion;
 import org.eclipse.jetty.http.MimeTypes;
 import org.eclipse.jetty.io.RuntimeIOException;
-import org.eclipse.jetty.server.Request;
 import org.eclipse.jetty.server.Response;
 import org.eclipse.jetty.server.handler.ContextResponse;
-import org.eclipse.jetty.session.Session;
-import org.eclipse.jetty.session.SessionManager;
-import org.eclipse.jetty.util.Blocker;
 import org.eclipse.jetty.util.Callback;
-import org.eclipse.jetty.util.FutureCallback;
 import org.eclipse.jetty.util.StringUtil;
-import org.eclipse.jetty.util.URIUtil;
 
 public class ServletContextResponse extends ContextResponse
 {
-    private static final int __MIN_BUFFER_SIZE = 1;
-    private static final EnumSet<EncodingFrom> __localeOverride = EnumSet.of(EncodingFrom.NOT_SET, EncodingFrom.DEFAULT, EncodingFrom.INFERRED, EncodingFrom.SET_LOCALE);
-
-    public enum OutputType
+    protected enum OutputType
     {
         NONE, STREAM, WRITER
     }
@@ -108,7 +85,92 @@ public class ServletContextResponse extends ContextResponse
         _response = response;
         _httpOutput = new HttpOutput(response, servletChannel);
         _servletChannel = servletChannel;
-        _httpServletResponse = new ServletApiResponse(this);
+        _httpServletResponse = newServletApiResponse();
+    }
+
+    protected ResponseWriter getWriter()
+    {
+        return _writer;
+    }
+
+    protected void setWriter(ResponseWriter writer)
+    {
+        _writer = writer;
+    }
+
+    protected Locale getLocale()
+    {
+        return _locale;
+    }
+
+    protected void setLocale(Locale locale)
+    {
+        _locale = locale;
+    }
+
+    protected EncodingFrom getEncodingFrom()
+    {
+        return _encodingFrom;
+    }
+
+    protected MimeTypes.Type getMimeType()
+    {
+        return _mimeType;
+    }
+
+    protected void setMimeType(MimeTypes.Type mimeType)
+    {
+        this._mimeType = mimeType;
+    }
+
+    protected Supplier<Map<String, String>> getTrailers()
+    {
+        return _trailers;
+    }
+
+    public void setTrailers(Supplier<Map<String, String>> trailers)
+    {
+        this._trailers = trailers;
+    }
+
+    protected void setContentType(String contentType)
+    {
+        this._contentType = contentType;
+    }
+
+    protected String getCharacterEncoding()
+    {
+        return _characterEncoding;
+    }
+
+    protected void setCharacterEncoding(String value)
+    {
+        _characterEncoding = value;
+    }
+
+    protected void setOutputType(OutputType outputType)
+    {
+        _outputType = outputType;
+    }
+
+    protected String getContentType()
+    {
+        return _contentType;
+    }
+
+    protected OutputType getOutputType()
+    {
+        return _outputType;
+    }
+
+    protected ServletContextRequest getServletContextRequest()
+    {
+        return _request;
+    }
+
+    protected ServletApiResponse newServletApiResponse()
+    {
+        return new ServletApiResponse(this);
     }
 
     public HttpOutput getHttpOutput()
@@ -163,6 +225,11 @@ public class ServletContextResponse extends ContextResponse
     }
 
     public void setContentLength(int len)
+    {
+        setContentLength((long)len);
+    }
+
+    public void setContentLength(long len)
     {
         // Protect from setting after committed as default handling
         // of a servlet HEAD request ALWAYS sets _content length, even
@@ -385,7 +452,7 @@ public class ServletContextResponse extends ContextResponse
         return encoding;
     }
 
-    private void setCharacterEncoding(String encoding, EncodingFrom from)
+    protected void setCharacterEncoding(String encoding, EncodingFrom from)
     {
         if (isWriting() || isCommitted())
             return;
@@ -445,7 +512,7 @@ public class ServletContextResponse extends ContextResponse
         return isWriting() || isStreaming();
     }
 
-    private enum EncodingFrom
+    protected enum EncodingFrom
     {
         /**
          * Character encoding was not set, or the encoding was cleared with {@code setCharacterEncoding(null)}.
@@ -476,612 +543,5 @@ public class ServletContextResponse extends ContextResponse
          * The character encoding has been explicitly set using {@link HttpServletResponse#setCharacterEncoding(String)}.
          */
         SET_CHARACTER_ENCODING
-    }
-
-    public class ServletApiResponse implements HttpServletResponse
-    {
-        private final ServletContextResponse _response;
-
-        ServletApiResponse(ServletContextResponse response)
-        {
-            _response = response;
-        }
-        
-        public ServletContextResponse getResponse()
-        {
-            return _response;
-        }
-
-        @Override
-        public void addCookie(Cookie cookie)
-        {
-            //Servlet Spec 9.3 Include method: cannot set a cookie if handling an include
-            if (StringUtil.isBlank(cookie.getName()))
-                throw new IllegalArgumentException("Cookie.name cannot be blank/null");
-
-            String comment = cookie.getComment();
-            // HttpOnly was supported as a comment in cookie flags before the java.net.HttpCookie implementation so need to check that
-            boolean httpOnlyFromComment = cookie.isHttpOnly() || HttpCookie.isHttpOnlyInComment(comment);
-            HttpCookie.SameSite sameSiteFromComment = HttpCookie.getSameSiteFromComment(comment);
-            comment = HttpCookie.getCommentWithoutAttributes(comment);
-            //old style cookie
-            if (sameSiteFromComment != null || httpOnlyFromComment)
-            {
-                addCookie(new HttpCookie(
-                    cookie.getName(),
-                    cookie.getValue(),
-                    cookie.getDomain(),
-                    cookie.getPath(),
-                    cookie.getMaxAge(),
-                    httpOnlyFromComment,
-                    cookie.getSecure(),
-                    comment,
-                    cookie.getVersion(),
-                    sameSiteFromComment));
-            }
-            else
-            {
-                //new style cookie, everything is an attribute
-                addCookie(new HttpCookie(
-                    cookie.getName(),
-                    cookie.getValue(),
-                    cookie.getVersion(),
-                    cookie.getAttributes()));
-            }
-        }
-
-        public void addCookie(HttpCookie cookie)
-        {
-            Response.addCookie(_response, cookie);
-        }
-
-        @Override
-        public boolean containsHeader(String name)
-        {
-            return _response.getHeaders().contains(name);
-        }
-
-        @Override
-        public String encodeURL(String url)
-        {
-            SessionManager sessionManager = _servletChannel.getContextHandler().getSessionHandler();
-            if (sessionManager == null)
-                return url;
-
-            ServletContextRequest request = _request;
-            HttpServletRequest httpServletRequest = request.getHttpServletRequest();
-
-            HttpURI uri = null;
-            if (sessionManager.isCheckingRemoteSessionIdEncoding() && URIUtil.hasScheme(url))
-            {
-                uri = HttpURI.from(url);
-                String path = uri.getPath();
-                path = (path == null ? "" : path);
-                int port = uri.getPort();
-                if (port < 0)
-                    port = HttpScheme.getDefaultPort(uri.getScheme());
-
-                // Is it the same server?
-                if (!Request.getServerName(request).equalsIgnoreCase(uri.getHost()))
-                    return url;
-                if (Request.getServerPort(request) != port)
-                    return url;
-                if (request.getContext() != null && !path.startsWith(request.getContext().getContextPath()))
-                    return url;
-            }
-
-            String sessionURLPrefix = sessionManager.getSessionIdPathParameterNamePrefix();
-            if (sessionURLPrefix == null)
-                return url;
-
-            if (url == null)
-                return null;
-
-            // should not encode if cookies in evidence
-            if ((sessionManager.isUsingCookies() && httpServletRequest.isRequestedSessionIdFromCookie()) || !sessionManager.isUsingURLs())
-            {
-                int prefix = url.indexOf(sessionURLPrefix);
-                if (prefix != -1)
-                {
-                    int suffix = url.indexOf("?", prefix);
-                    if (suffix < 0)
-                        suffix = url.indexOf("#", prefix);
-
-                    if (suffix <= prefix)
-                        return url.substring(0, prefix);
-                    return url.substring(0, prefix) + url.substring(suffix);
-                }
-                return url;
-            }
-
-            // get session;
-            HttpSession session = httpServletRequest.getSession(false);
-
-            // no session
-            if (session == null || !(session instanceof Session.APISession))
-                return url;
-
-            // invalid session
-            Session.APISession apiSession = (Session.APISession)session;
-
-            if (!apiSession.getCoreSession().isValid())
-                return url;
-
-            String id = apiSession.getCoreSession().getExtendedId();
-
-            if (uri == null)
-                uri = HttpURI.from(url);
-
-            // Already encoded
-            int prefix = url.indexOf(sessionURLPrefix);
-            if (prefix != -1)
-            {
-                int suffix = url.indexOf("?", prefix);
-                if (suffix < 0)
-                    suffix = url.indexOf("#", prefix);
-
-                if (suffix <= prefix)
-                    return url.substring(0, prefix + sessionURLPrefix.length()) + id;
-                return url.substring(0, prefix + sessionURLPrefix.length()) + id +
-                    url.substring(suffix);
-            }
-
-            // edit the session
-            int suffix = url.indexOf('?');
-            if (suffix < 0)
-                suffix = url.indexOf('#');
-            if (suffix < 0)
-            {
-                return url +
-                    ((HttpScheme.HTTPS.is(uri.getScheme()) || HttpScheme.HTTP.is(uri.getScheme())) && uri.getPath() == null ? "/" : "") + //if no path, insert the root path
-                    sessionURLPrefix + id;
-            }
-
-            return url.substring(0, suffix) +
-                ((HttpScheme.HTTPS.is(uri.getScheme()) || HttpScheme.HTTP.is(uri.getScheme())) && uri.getPath() == null ? "/" : "") + //if no path so insert the root path
-                sessionURLPrefix + id + url.substring(suffix);
-        }
-
-        @Override
-        public String encodeRedirectURL(String url)
-        {
-            return encodeURL(url);
-        }
-
-        @Override
-        public void sendError(int sc, String msg) throws IOException
-        {
-            switch (sc)
-            {
-                case -1 -> _servletChannel.abort(new IOException(msg));
-                case HttpStatus.PROCESSING_102, HttpStatus.EARLY_HINT_103 ->
-                {
-                    if (!isCommitted())
-                    {
-                        try (Blocker.Callback blocker = Blocker.callback())
-                        {
-                            CompletableFuture<Void> completable = _response.writeInterim(sc, _response.getHeaders().asImmutable());
-                            blocker.completeWith(completable);
-                            blocker.block();
-                        }
-                    }
-                }
-                default -> getState().sendError(sc, msg);
-            }
-        }
-
-        @Override
-        public void sendError(int sc) throws IOException
-        {
-            sendError(sc, null);
-        }
-
-        @Override
-        public void sendRedirect(String location) throws IOException
-        {
-            sendRedirect(HttpServletResponse.SC_MOVED_TEMPORARILY, location);
-        }
-
-        /**
-         * Sends a response with one of the 300 series redirection codes.
-         *
-         * @param code the redirect status code
-         * @param location the location to send in {@code Location} headers
-         * @throws IOException if unable to send the redirect
-         */
-        public void sendRedirect(int code, String location) throws IOException
-        {
-            resetBuffer();
-            FutureCallback callback = new FutureCallback();
-            Response.sendRedirect(_request, _response, callback, code, location, false);
-            callback.block();
-        }
-
-        @Override
-        public void setDateHeader(String name, long date)
-        {
-            _response.getHeaders().putDateField(name, date);
-        }
-
-        @Override
-        public void addDateHeader(String name, long date)
-        {
-            _response.getHeaders().addDateField(name, date);
-        }
-
-        @Override
-        public void setHeader(String name, String value)
-        {
-            _response.getHeaders().put(name, value);
-        }
-
-        @Override
-        public void addHeader(String name, String value)
-        {
-            _response.getHeaders().add(name, value);
-        }
-
-        @Override
-        public void setIntHeader(String name, int value)
-        {
-            // TODO do we need int versions?
-            _response.getHeaders().putLongField(name, value);
-        }
-
-        @Override
-        public void addIntHeader(String name, int value)
-        {
-            // TODO do we need a native version?
-            _response.getHeaders().add(name, Integer.toString(value));
-        }
-
-        @Override
-        public void setStatus(int sc)
-        {
-            _response.setStatus(sc);
-        }
-
-        @Override
-        public int getStatus()
-        {
-            return _response.getStatus();
-        }
-
-        @Override
-        public String getHeader(String name)
-        {
-            return _response.getHeaders().get(name);
-        }
-
-        @Override
-        public Collection<String> getHeaders(String name)
-        {
-            return _response.getHeaders().getValuesList(name);
-        }
-
-        @Override
-        public Collection<String> getHeaderNames()
-        {
-            return _response.getHeaders().getFieldNamesCollection();
-        }
-
-        @Override
-        public String getCharacterEncoding()
-        {
-            return _response.getCharacterEncoding(false);
-        }
-
-        @Override
-        public String getContentType()
-        {
-            return _contentType;
-        }
-
-        @Override
-        public ServletOutputStream getOutputStream() throws IOException
-        {
-            if (_outputType == OutputType.WRITER)
-                throw new IllegalStateException("WRITER");
-            _outputType = OutputType.STREAM;
-            return _httpOutput;
-        }
-
-        @Override
-        public PrintWriter getWriter() throws IOException
-        {
-            if (_outputType == OutputType.STREAM)
-                throw new IllegalStateException("STREAM");
-
-            if (_outputType == OutputType.NONE)
-            {
-                String encoding = _response.getCharacterEncoding(true);
-                Locale locale = getLocale();
-                if (_writer != null && _writer.isFor(locale, encoding))
-                    _writer.reopen();
-                else
-                {
-                    if (StringUtil.__ISO_8859_1.equalsIgnoreCase(encoding))
-                        _writer = new ResponseWriter(new Iso88591HttpWriter(_httpOutput), locale, encoding);
-                    else if (StringUtil.__UTF8.equalsIgnoreCase(encoding))
-                        _writer = new ResponseWriter(new Utf8HttpWriter(_httpOutput), locale, encoding);
-                    else
-                        _writer = new ResponseWriter(new EncodingHttpWriter(_httpOutput, encoding), locale, encoding);
-                }
-
-                // Set the output type at the end, because setCharacterEncoding() checks for it.
-                _outputType = OutputType.WRITER;
-            }
-            return _writer;
-        }
-
-        @Override
-        public void setCharacterEncoding(String encoding)
-        {
-            _response.setCharacterEncoding(encoding, EncodingFrom.SET_CHARACTER_ENCODING);
-        }
-
-        @Override
-        public void setContentLength(int len)
-        {
-            // Protect from setting after committed as default handling
-            // of a servlet HEAD request ALWAYS sets _content length, even
-            // if the getHandling committed the response!
-            if (isCommitted())
-                return;
-
-            if (len > 0)
-            {
-                long written = _httpOutput.getWritten();
-                if (written > len)
-                    throw new IllegalArgumentException("setContentLength(" + len + ") when already written " + written);
-
-                _contentLength = len;
-                _response.getHeaders().putLongField(HttpHeader.CONTENT_LENGTH, len);
-                if (isAllContentWritten(written))
-                {
-                    try
-                    {
-                        closeOutput();
-                    }
-                    catch (IOException e)
-                    {
-                        throw new RuntimeIOException(e);
-                    }
-                }
-            }
-            else if (len == 0)
-            {
-                long written = _httpOutput.getWritten();
-                if (written > 0)
-                    throw new IllegalArgumentException("setContentLength(0) when already written " + written);
-                _contentLength = len;
-                _response.getHeaders().put(HttpHeader.CONTENT_LENGTH, "0");
-            }
-            else
-            {
-                _contentLength = len;
-                _response.getHeaders().remove(HttpHeader.CONTENT_LENGTH);
-            }
-        }
-
-        @Override
-        public void setContentLengthLong(long len)
-        {
-            // Protect from setting after committed as default handling
-            // of a servlet HEAD request ALWAYS sets _content length, even
-            // if the getHandling committed the response!
-            if (isCommitted())
-                return;
-            _contentLength = len;
-            _response.getHeaders().putLongField(HttpHeader.CONTENT_LENGTH.toString(), len);
-        }
-
-        @Override
-        public void setContentType(String contentType)
-        {
-            if (isCommitted())
-                return;
-
-            if (contentType == null)
-            {
-                if (isWriting() && _characterEncoding != null)
-                    throw new IllegalSelectorException();
-
-                if (_locale == null)
-                    _characterEncoding = null;
-                _mimeType = null;
-                _contentType = null;
-                _response.getHeaders().remove(HttpHeader.CONTENT_TYPE);
-            }
-            else
-            {
-                _contentType = contentType;
-                _mimeType = MimeTypes.CACHE.get(contentType);
-
-                String charset = MimeTypes.getCharsetFromContentType(contentType);
-                if (charset == null && _mimeType != null && _mimeType.isCharsetAssumed())
-                    charset = _mimeType.getCharsetString();
-
-                if (charset == null)
-                {
-                    switch (_encodingFrom)
-                    {
-                        case NOT_SET:
-                            break;
-                        case DEFAULT:
-                        case INFERRED:
-                        case SET_CONTENT_TYPE:
-                        case SET_LOCALE:
-                        case SET_CHARACTER_ENCODING:
-                        {
-                            _contentType = contentType + ";charset=" + _characterEncoding;
-                            _mimeType = MimeTypes.CACHE.get(_contentType);
-                            break;
-                        }
-                        default:
-                            throw new IllegalStateException(_encodingFrom.toString());
-                    }
-                }
-                else if (isWriting() && !charset.equalsIgnoreCase(_characterEncoding))
-                {
-                    // too late to change the character encoding;
-                    _contentType = MimeTypes.getContentTypeWithoutCharset(_contentType);
-                    if (_characterEncoding != null  && (_mimeType == null || !_mimeType.isCharsetAssumed()))
-                        _contentType = _contentType + ";charset=" + _characterEncoding;
-                    _mimeType = MimeTypes.CACHE.get(_contentType);
-                }
-                else
-                {
-                    _characterEncoding = charset;
-                    _encodingFrom = EncodingFrom.SET_CONTENT_TYPE;
-                }
-
-                if (HttpGenerator.__STRICT || _mimeType == null)
-                    _response.getHeaders().put(HttpHeader.CONTENT_TYPE, _contentType);
-                else
-                {
-                    _contentType = _mimeType.asString();
-                    _response.getHeaders().put(_mimeType.getContentTypeField());
-                }
-            }
-        }
-
-        public long getContentCount()
-        {
-            return _httpOutput.getWritten();
-        }
-
-        @Override
-        public void setBufferSize(int size)
-        {
-            if (isCommitted())
-                throw new IllegalStateException("cannot set buffer size after response is in committed state");
-            if (getContentCount() > 0)
-                throw new IllegalStateException("cannot set buffer size after response has " + getContentCount() + " bytes already written");
-            if (size < __MIN_BUFFER_SIZE)
-                size = __MIN_BUFFER_SIZE;
-            _httpOutput.setBufferSize(size);
-        }
-
-        @Override
-        public int getBufferSize()
-        {
-            return _httpOutput.getBufferSize();
-        }
-
-        @Override
-        public void flushBuffer() throws IOException
-        {
-            if (!_httpOutput.isClosed())
-                _httpOutput.flush();
-        }
-
-        @Override
-        public void resetBuffer()
-        {
-            _httpOutput.resetBuffer();
-            _httpOutput.reopen();
-        }
-
-        @Override
-        public boolean isCommitted()
-        {
-            // If we are in sendError state, we pretend to be committed
-            if (_servletChannel.isSendError())
-                return true;
-            return _servletChannel.isCommitted();
-        }
-
-        @Override
-        public void reset()
-        {
-            if (isCommitted())
-                throw new IllegalStateException("Committed");
-
-            _response.reset();
-
-            ServletApiRequest servletApiRequest = _request.getServletApiRequest();
-            Session session = servletApiRequest.getCoreSession();
-            if (session != null && session.isNew())
-            {
-                SessionManager sessionManager = servletApiRequest.getSessionManager();
-                if (sessionManager != null)
-                {
-                    HttpCookie cookie = sessionManager.getSessionCookie(session, servletApiRequest.getContextPath(), servletApiRequest.getServletConnection().isSecure());
-                    if (cookie != null)
-                        addCookie(cookie);
-                }
-            }
-        }
-
-        @Override
-        public void setLocale(Locale locale)
-        {
-            if (isCommitted())
-                return;
-
-            if (locale == null)
-            {
-                _locale = null;
-                _response.getHeaders().remove(HttpHeader.CONTENT_LANGUAGE);
-                if (_encodingFrom == EncodingFrom.SET_LOCALE)
-                    _response.setCharacterEncoding(null, EncodingFrom.NOT_SET);
-            }
-            else
-            {
-                _locale = locale;
-                _response.getHeaders().put(HttpHeader.CONTENT_LANGUAGE, StringUtil.replace(locale.toString(), '_', '-'));
-
-                if (_outputType != OutputType.NONE)
-                    return;
-
-                ServletContextHandler.ServletScopedContext context = _servletChannel.getContext();
-                if (context == null)
-                    return;
-
-                String charset = context.getServletContextHandler().getLocaleEncoding(locale);
-                if (!StringUtil.isEmpty(charset) && __localeOverride.contains(_encodingFrom))
-                    _response.setCharacterEncoding(charset, EncodingFrom.SET_LOCALE);
-            }
-        }
-
-        @Override
-        public Locale getLocale()
-        {
-            if (_locale == null)
-                return Locale.getDefault();
-            return _locale;
-        }
-
-        @Override
-        public Supplier<Map<String, String>> getTrailerFields()
-        {
-            return _trailers;
-        }
-
-        @Override
-        public void setTrailerFields(Supplier<Map<String, String>> trailers)
-        {
-            if (isCommitted())
-                throw new IllegalStateException("Committed");
-            HttpVersion version = HttpVersion.fromString(_request.getConnectionMetaData().getProtocol());
-            if (version == null || version.compareTo(HttpVersion.HTTP_1_1) < 0)
-                throw new IllegalStateException("Trailers not supported in " + version);
-
-            _trailers = trailers;
-
-            _response.setTrailersSupplier(() ->
-            {
-                Map<String, String> map = trailers.get();
-                if (map == null)
-                    return null;
-                HttpFields.Mutable fields = HttpFields.build(map.size());
-                for (Map.Entry<String, String> e : map.entrySet())
-                {
-                    fields.add(e.getKey(), e.getValue());
-                }
-                return fields;
-            });
-        }
     }
 }
