+++ jetty-ee10/jetty-ee10-servlet/src/main/java/org/eclipse/jetty/ee10/servlet/security/ConstraintSecurityHandler.java
@@ -33,6 +33,7 @@ import jakarta.servlet.annotation.ServletSecurity.EmptyRoleSemantic;
 import jakarta.servlet.annotation.ServletSecurity.TransportGuarantee;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.eclipse.jetty.ee10.servlet.ServletApiRequest;
 import org.eclipse.jetty.ee10.servlet.ServletContextHandler;
 import org.eclipse.jetty.ee10.servlet.ServletContextRequest;
 import org.eclipse.jetty.http.HttpHeader;
@@ -681,7 +682,7 @@ public class ConstraintSecurityHandler extends SecurityHandler implements Constr
         throws IOException
     {
         ServletContextRequest screquest = Request.as(request, ServletContextRequest.class);
-        ServletContextRequest.ServletApiRequest sarequest = (screquest == null ? null : screquest.getServletApiRequest());
+        ServletApiRequest sarequest = (screquest == null ? null : screquest.getServletApiRequest());
         
         if (constraintInfo == null)
         {
