+++ jetty-ee10/jetty-ee10-servlet/src/test/java/org/eclipse/jetty/ee10/servlet/ServletContextHandlerTest.java
@@ -51,6 +51,7 @@ import jakarta.servlet.ServletRequestEvent;
 import jakarta.servlet.ServletRequestListener;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.SessionTrackingMode;
+import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -64,7 +65,11 @@ import org.eclipse.jetty.ee10.servlet.security.ConstraintSecurityHandler;
 import org.eclipse.jetty.ee10.servlet.security.RoleInfo;
 import org.eclipse.jetty.ee10.servlet.security.SecurityHandler;
 import org.eclipse.jetty.ee10.servlet.security.UserIdentity;
+import org.eclipse.jetty.http.HttpCookie;
+import org.eclipse.jetty.http.HttpHeader;
 import org.eclipse.jetty.http.HttpTester;
+import org.eclipse.jetty.http.pathmap.MatchedPath;
+import org.eclipse.jetty.http.pathmap.PathSpec;
 import org.eclipse.jetty.logging.StacklessLogging;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.LocalConnector;
@@ -2367,5 +2372,100 @@ public class ServletContextHandlerTest
         response = _connector.getResponse(request);
         assertThat(response, containsString("200 OK"));
         assertThat(response, containsString("/three"));
+    }
+
+    static class CookieTweakResponseApi extends ServletApiResponse
+    {
+        CookieTweakResponseApi(ServletApiResponse response)
+        {
+            super(response.getResponse());
+        }
+
+        @Override
+        public void addCookie(Cookie cookie)
+        {
+            // Let's set SameSite to STRICT
+            // But we cannot use Servlet jakarta.servlet.http.Cookie.setComment(String) technique of old
+            // as the Comment field is always null and is slated for removal.
+
+            // We'll use the Jetty HttpCookie to work around this.
+            boolean httpOnly = false;
+            String comment = null;
+            HttpCookie.SameSite sameSite = HttpCookie.SameSite.STRICT;
+
+            addCookie(new HttpCookie(
+                cookie.getName(),
+                cookie.getValue(),
+                cookie.getDomain(),
+                cookie.getPath(),
+                cookie.getMaxAge(),
+                httpOnly,
+                cookie.getSecure(),
+                comment,
+                cookie.getVersion(),
+                HttpCookie.SameSite.STRICT));
+        }
+    }
+
+    @Test
+    public void testCustomServletResponseWrapping() throws Exception
+    {
+        ServletContextHandler context = new ServletContextHandler()
+        {
+            @Override
+            protected ServletContextRequest newServletContextRequest(ServletChannel servletChannel, Request request, Response response, String pathInContext, ServletHandler.MappedServlet mappedServlet, PathSpec pathSpec, MatchedPath matchedPath)
+            {
+                return new ServletContextRequest(getContext().getServletContext(), servletChannel,
+                    request, response, pathInContext,
+                    mappedServlet, pathSpec, matchedPath) {
+
+                    @Override
+                    protected ServletApiRequest newServletApiRequest()
+                    {
+                        return super.newServletApiRequest();
+                    }
+
+                    @Override
+                    protected ServletContextResponse newServletContextResponse()
+                    {
+                        return new ServletContextResponse(servletChannel, this, response)
+                        {
+                            @Override
+                            protected ServletApiResponse newServletApiResponse()
+                            {
+                                ServletApiResponse servletApiResponse = super.newServletApiResponse();
+                                return new CookieTweakResponseApi(super.newServletApiResponse());
+                            }
+                        };
+                    }
+                };
+            }
+        };
+
+        ServletHolder holder = new ServletHolder(new HttpServlet()
+        {
+            @Override
+            protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException
+            {
+                Cookie cookie = new Cookie("example", "bogus");
+                resp.addCookie(cookie);
+                super.doGet(req, resp);
+            }
+        });
+        context.addServlet(holder, "/cookies/*");
+        _server.setHandler(context);
+        _server.start();
+
+        String rawRequest = """
+            GET /cookies/ HTTP/1.1\r
+            Host: localhost\r
+            Connection: close\r
+            \r
+            """;
+
+        String rawResponse = _connector.getResponse(rawRequest);
+        HttpTester.Response response = HttpTester.parseResponse(rawResponse);
+        String setCookieValue = response.get(HttpHeader.SET_COOKIE);
+        assertThat(setCookieValue, containsString("example=bogus; SameSite=Strict"));
     }
 }
