+++ jetty-ee10/test-ee10-sessions/test-ee10-sessions-common/src/test/java/org/eclipse/jetty/ee10/session/IdleSessionTest.java
@@ -24,8 +24,8 @@ import jakarta.servlet.http.HttpSession;
 import org.eclipse.jetty.client.ContentResponse;
 import org.eclipse.jetty.client.HttpClient;
 import org.eclipse.jetty.client.Request;
+import org.eclipse.jetty.ee10.servlet.ServletApiRequest;
 import org.eclipse.jetty.ee10.servlet.ServletContextHandler;
-import org.eclipse.jetty.ee10.servlet.ServletContextRequest;
 import org.eclipse.jetty.ee10.servlet.ServletHolder;
 import org.eclipse.jetty.logging.StacklessLogging;
 import org.eclipse.jetty.session.DefaultSessionCacheFactory;
@@ -264,7 +264,7 @@ public class IdleSessionTest
                 session.setAttribute("value", 1);
                 originalId = session.getId();
 
-                Session s = ServletContextRequest.ServletApiRequest.getSession(session);
+                Session s = ServletApiRequest.getSession(session);
                 try (AutoLock lock = s.lock())
                 {
                     assertTrue(s.isResident());
@@ -276,7 +276,7 @@ public class IdleSessionTest
                 HttpSession session = request.getSession(false);
                 assertNotNull(session);
                 assertEquals(originalId, session.getId());
-                Session s = ServletContextRequest.ServletApiRequest.getSession(session);
+                Session s = ServletApiRequest.getSession(session);
                 try (AutoLock lock = s.lock())
                 {
                     assertTrue(s.isResident());
